backend/base/views/user_views.py
#Django Import
from django.core.exceptions import RequestDataTooBig
from django.shortcuts import render
from datetime import datetime


#Rest framework Import
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated, IsAdminUser
from rest_framework.response import Response
from rest_framework.serializers import Serializer

#Local Import

